Update StockBatch DTOs and service logic: - CreateStockBatch and UpdateStockBatch SQL DTOs for handling stock batch creation and updates - Add business logic for creating a new stock batch - Ensure the business logic covers validation and data consistency rules

+++ internal/entity/dtos/sql/stockbatch/create_stockbatch.go
@@ -5,11 +5,11 @@ import (
 )
 
 type CreateStockBatch struct {
-	StockBatchID   uint              `gorm:"column:stockbatch_id;primaryKey;autoIncrement" json:"-"`
+	StockBatchID   uint64            `gorm:"column:stockbatch_id;primaryKey;autoIncrement" json:"-"`
 	Quantity       int               `gorm:"column:quantity;not null" json:"quantity" validate:"required,gt=0"`
 	ExpirationDate common.CustomDate `gorm:"column:expiration_date;not null" json:"expiration_date" validate:"required"`
 	ReceivedDate   common.CustomDate `gorm:"column:received_date;not null" json:"received_date" validate:"required"`
-	IngredientID   uint              `gorm:"column:ingredient_id;not null" json:"ingredient_id" validate:"required"`
+	IngredientID   uint64            `gorm:"column:ingredient_id;not null" json:"ingredient_id" validate:"required"`
 	common.CommonFields
 }
